+++ pages/calendario.php
@@ -10,7 +10,7 @@ if (!isset($_SESSION['usuario'])){
 $page_title="Calendario";
 $user=$_SESSION['usuario'];
 //Get user query
-$persona = mysqli_query($con, "SELECT * FROM persona WHERE email='$user' AND borrado = 0");
+$persona = mysqli_query($con, "SELECT * FROM persona WHERE email='$user' and borrado=0");
 $rowp = mysqli_fetch_assoc($persona);
 $id_rowp = $rowp['id_persona'];